+++ apps/meteor/client/views/room/composer/messageBox/MessageBox.tsx
@@ -83,7 +83,7 @@ type MessageBoxProps = {
 	rid: IRoom['_id'];
 	tmid?: IMessage['_id'];
 	readOnly: boolean;
-	onSend?: (params: { value: string; tshow?: boolean }) => Promise<void>;
+	onSend?: (params: { value: string; tshow?: boolean; previewUrls?: string[] }) => Promise<void>;
 	onJoin?: () => Promise<void>;
 	onResize?: () => void;
 	onTyping?: () => void;
@@ -92,6 +92,7 @@ type MessageBoxProps = {
 	onNavigateToNextMessage?: () => void;
 	onUploadFiles?: (files: readonly File[]) => void;
 	tshow?: IMessage['tshow'];
+	previewUrls?: string[];
 	subscription?: ISubscription;
 	showFormattingTips: boolean;
 	isEmbedded?: boolean;
@@ -109,6 +110,7 @@ const MessageBox = ({
 	onTyping,
 	readOnly,
 	tshow,
+	previewUrls,
 }: MessageBoxProps): ReactElement => {
 	const chat = useChat();
 	const t = useTranslation();
@@ -203,6 +205,7 @@ const MessageBox = ({
 		onSend?.({
 			value: text,
 			tshow,
+			previewUrls,
 		});
 	});