+++ apps/meteor/ee/server/apps/orchestrator.js
@@ -15,6 +15,7 @@ import {
 	AppDepartmentsConverter,
 	AppUploadsConverter,
 	AppVisitorsConverter,
+	AppRolesConverter,
 } from '../../../app/apps/server/converters';
 import { AppThreadsConverter } from '../../../app/apps/server/converters/threads';
 import { settings, settingsRegistry } from '../../../app/settings/server';
@@ -25,6 +26,8 @@ import { AppRealLogsStorage, AppRealStorage, ConfigurableAppSourceStorage } from
 function isTesting() {
 	return process.env.TEST_MODE === 'true';
 }
+
+const DISABLED_PRIVATE_APP_INSTALLATION = ['yes', 'true'].includes(String(process.env.DISABLE_PRIVATE_APP_INSTALLATION).toLowerCase());
 
 let appsSourceStorageType;
 let appsSourceStorageFilesystemPath;
@@ -64,6 +67,7 @@ export class AppServerOrchestrator {
 		this._converters.set('uploads', new AppUploadsConverter(this));
 		this._converters.set('videoConferences', new AppVideoConferencesConverter());
 		this._converters.set('threads', new AppThreadsConverter(this));
+		this._converters.set('roles', new AppRolesConverter(this));
 
 		this._bridges = new RealAppBridges(this);
 
@@ -135,6 +139,10 @@ export class AppServerOrchestrator {
 
 	isDebugging() {
 		return !isTesting();
+	}
+
+	shouldDisablePrivateAppInstallation() {
+		return DISABLED_PRIVATE_APP_INSTALLATION;
 	}
 
 	/**