+++ packages/rest-typings/src/v1/users/UsersSetPreferenceParamsPOST.ts
@@ -50,6 +50,8 @@ export type UsersSetPreferencesParamsPOST = {
 		idleTimeLimit?: number;
 		omnichannelTranscriptEmail?: boolean;
 		omnichannelTranscriptPDF?: boolean;
+		enableMobileRinging?: boolean;
+		mentionsWithSymbol?: boolean;
 	};
 };
 
@@ -242,6 +244,14 @@ const UsersSetPreferencesParamsPostSchema = {
 					nullable: true,
 				},
 				omnichannelTranscriptPDF: {
+					type: 'boolean',
+					nullable: true,
+				},
+				enableMobileRinging: {
+					type: 'boolean',
+					nullable: true,
+				},
+				mentionsWithSymbol: {
 					type: 'boolean',
 					nullable: true,
 				},