+++ packages/ui-video-conf/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @rocket.chat/ui-video-conf
 
+## 1.0.4
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@1.0.1
+
 ## 1.0.0
 
 ### Patch Changes